Epic quest to fix cursor jiggle
Had to write a Render routine, keep track of virtual dom nodes on routes
and swap in/out when needed.

Switched to using an ::after pseudo-element on the actively typed
element for the cursor because when using in-memory DOM nodes the
browser will not give you the rendered coordinates of the node.