Change Render() to mount the actual route DOM object instead of cloning
Cloning was _SUPER_ slow.  So much so that it ended up blocking the main
thread for long enough for click callbacks to not fire.  Obviously, this
property is a non-starter for that approach.

Furthermore the only reason I implemented that was to get around
cursor-flicker, which seems to have been weeded out by changing the
cursor to an ::after pseudo-element on the actively typed thing, so yay!